Move util files to individual packages
