+++ pkg/kubernetes/types/pods.go
@@ -1,14 +1,17 @@
 package types
 
-import "github.com/ovn-kubernetes/ovn-kubernetes-mcp/pkg/utils"
+import (
+	"github.com/ovn-kubernetes/ovn-kubernetes-mcp/pkg/utils/headtail"
+	"github.com/ovn-kubernetes/ovn-kubernetes-mcp/pkg/utils/pattern"
+)
 
 // GetPodLogsParams is a type that contains the name, namespace and container of a pod.
 type GetPodLogsParams struct {
 	NamespacedNameParams
 	Container string `json:"container,omitempty"`
 	Previous  bool   `json:"previous,omitempty"`
-	utils.PatternParams
-	utils.HeadTailParams
+	pattern.PatternParams
+	headtail.HeadTailParams
 }
 
 // GetPodLogsResult is a type that contains the logs of a pod where each log line
