+++ pkg/utils/commandbuilder/command_builder.go
@@ -1,4 +1,4 @@
-package utils
+package commandbuilder
 
 import "slices"
