+++ pkg/utils/headtail/head_tail_test.go
@@ -1,4 +1,4 @@
-package utils
+package headtail
 
 import (
 	"slices"
