+++ pkg/utils/ovndb/ovn_db_test.go
@@ -1,4 +1,4 @@
-package utils
+package ovndb
 
 import "testing"
