+++ pkg/utils/pattern/pattern.go
@@ -1,4 +1,4 @@
-package utils
+package pattern
 
 import (
 	"fmt"
