+++ pkg/utils/pattern/pattern_test.go
@@ -1,4 +1,4 @@
-package utils
+package pattern
 
 import (
 	"slices"
